Added emails controller
