+++ app/controllers/emails_controller.rb
@@ -1,0 +1,2 @@
+class EmailsController < ApiController
+end
