+++ spec/controllers/emails_controller_spec.rb
@@ -1,0 +1,9 @@
+require 'spec_helper'
+
+describe EmailsController do
+
+  it "should be an api controller" do
+    EmailsController.new.should be_a(ApiController)
+  end
+
+end
